requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM clients
anthropic>=0.7.0

# NLP and embeddings
spacy>=3.7.0
transformers>=4.35.0
# torch will be installed as a dependency of transformers/sentence-transformers
# If you need CPU-only torch: pip install torch --index-url https://download.pytorch.org/whl/cpu
sentence-transformers>=2.2.0

# Vector database
chromadb>=0.4.15

# Data processing
pandas>=2.1.0
numpy>=1.24.0,<2.0.0  # NumPy 2.x incompatible with PyTorch

# Web utilities (optional - not required for core functionality)
# beautifulsoup4>=4.12.0  # Only needed if scraping
# requests>=2.31.0  # Only needed if scraping

# Utilities
tqdm>=4.66.0

